requirements.txt
catboost
category_encoders
ctgan
icecream
imbalanced_learn==0.12.3
lightgbm
matplotlib
ml_collections
numpy
pandas
peft
polars
prdc
PyYAML
rtdl_num_embeddings
scikit_learn
scipy
sdmetrics
seaborn
tabm
tomli
tomli_w
torch
torch_geometric
torch_sparse
tqdm
transformers
xgboost==2.1.3
zero